# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DDAS API",
    "version": "2.0.0-alpha.0",
    "description": "Dynamic Delegation of Authority System — appetite-constrained authority routing for humans and AI agents."
  },
...
  "paths": {
...
    "/api/v1/approval-tasks/{id}": {
      "get": {
        "tags": [
          "approvals"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "classificationId": {
                      "type": "string"
                    },
                    "requiredTier": {
                      "type": "number"
                    },
                    "quorum": {
                      "type": "number"
                    },
                    "dueAt": {
                      "type": "string"
                    },
                    "escalationLevel": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "open",
                        "decided",
                        "failed"
                      ]
                    },
                    "resolutionTrace": {},
                    "approvers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "principalId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "via": {
                            "type": "string",
                            "enum": [
                              "position",
                              "delegation",
                              "escalation"
                            ]
                          }
                        },
                        "required": [
                          "principalId",
                          "name",
                          "via"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "actions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "principalId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "action": {
                            "type": "string",
                            "enum": [
                              "approve",
                              "reject"
                            ]
                          },
                          "comment": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "principalId",
                          "name",
                          "action",
                          "comment",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "requestId",
                    "classificationId",
                    "requiredTier",
                    "quorum",
                    "dueAt",
                    "escalationLevel",
                    "status",
                    "approvers",
                    "actions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approval-tasks/{id}/approve": {
      "post": {
        "tags": [
          "approvals"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment": {
                    "type": "string"
                  }
                },
                "additionalProperties": false,
                "default": {}
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "verdict": {
                      "type": "string",
                      "enum": [
                        "approved",
                        "rejected",
                        "pending"
                      ]
                    },
                    "approvals": {
                      "type": "number"
                    },
                    "quorum": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "verdict",
                    "approvals",
                    "quorum"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approval-tasks/{id}/reject": {
      "post": {
        "tags": [
          "approvals"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "comment"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "verdict": {
                      "type": "string",
                      "enum": [
                        "approved",
                        "rejected",
                        "pending"
                      ]
                    },
                    "approvals": {
                      "type": "number"
                    },
                    "quorum": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "verdict",
                    "approvals",
                    "quorum"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}